Player Shooting & Started on enemy AI
Player now shoots a bullet in the direction of the mouse position when clicking LMB. Enemies will be destroyed after being hit by these bullets x amount of times. When the bullets are out of view they will be destroyed so they don't lag the game.

Started on the basic enemy AI, made attempts to get the enemy to lock on to the player and follow them but it's very buggy at the moment. Enemy starts moving but can't seem to turn, the movement is too floaty.